requirements.txt
absl-py==0.11.0
alabaster==0.7.12
ants==0.0.7
argon2-cffi==20.1.0
asgiref==3.3.1
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
autograd==1.3
Babel==2.9.0
backcall==0.2.0
backports.functools-lru-cache==1.6.1
bleach==3.2.1
blosc==1.10.2
cached-property==1.5.2
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==8.0.3
cycler==0.10.0
Cython==3.0a6
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.6
docutils==0.16
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gdown==3.12.2
gevent==21.1.2
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
greenlet==1.0.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
imagesize==1.2.0
importlib-metadata==3.3.0
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itk==5.2rc1
itk-core==5.2rc1
itk-filtering==5.2rc1
itk-io==5.2rc1
itk-numerics==5.2rc1
itk-registration==5.2rc1
itk-segmentation==5.2rc1
jax==0.2.10
jaxlib==0.1.61
jdcal==1.4.1
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lagomorph==1.0.25
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mermaid==v0.2.4-20-gcb48c72
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nbsphinx==0.8.1
nest-asyncio==1.4.3
networkx==2.5
neurite==0.1
nibabel==3.2.1
notebook==6.1.6
numpy==1.19.5
oauthlib==3.1.0
openpyxl==3.0.6
opt-einsum==3.3.0
packaging==20.8
pandas==1.2.0
pandoc==2.0a4
pandocfilters==1.4.3
parso==0.8.1
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pip==20.3.3
plumbum==1.6.9
ply==3.11
progressbar2==3.53.1
prometheus-client==0.9.0
prompt-toolkit==3.0.10
protobuf==3.14.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.3
pynrrd==0.4.2
pypandoc==1.5
pyparsing==3.0.0b2
pyrsistent==0.17.3
PySocks==1.7.1
pystrum==0.1
python-dateutil==2.8.1
python-utils==2.5.6
pytz==2020.5
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==20.0.0
qtconsole==5.0.2
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.18.1
scikit-learn==0.24.0
scipy==1.6.0
seaborn==0.11.2
Send2Trash==1.5.0
setuptools==49.6.0.post20210108
SimpleITK==2.0.2
six==1.15.0
sklearn==0.0
snowballstemmer==2.1.0
Sphinx==3.4.3
sphinx-gallery==0.3.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
statsmodels==0.12.2
tabulate==0.8.7
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2021.1.8
torch==1.10.0+cu113
torchaudio==0.10.0+cu113
torchvision==0.11.1+cu113
tornado==6.1
tqdm==4.55.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.2
voxelmorph==0.1
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
Werkzeug==1.0.1
wheel==0.36.2
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0
